code/requirements.txt
numpy==1.16.2
tensorflow-gpu==2.0.0
tensorflow_model_optimization
neural_structured_learning==1.1.0
matplotlib
absl_py
PIL
opencv
